+++ bot/_logger.py
@@ -1,6 +1,6 @@
 import logging
 
-from .user import get_logfile
+from .util import get_logfile
 
 formatter = logging.Formatter('%(module)s : %(levelname)s : %(asctime)s : %(message)s')
