+++ bot/alert_processes/__init__.py
@@ -1,4 +1,4 @@
-# Import for consolidated use in .alerting import
+# Import for consolidated use in .alert_processes import
 from .base import BaseAlertProcess
 from .cex import CEXAlertProcess
 from .dex import DEXAlertProcess
