+++ bot/alert_processes/technical.py
@@ -4,7 +4,7 @@ import os
 from typing import Union
 
 from .base import BaseAlertProcess
-from ..user import LocalUserConfiguration, MongoDBUserConfiguration, get_whitelist
+from ..user_configuration import LocalUserConfiguration, MongoDBUserConfiguration, get_whitelist
 from .._logger import logger
 from ..config import *
 from ..indicators import TADatabaseClient, TAAggregateClient
